+++ CHANGELOG.md
@@ -1,10 +1,18 @@
 # CHANGELOG
 
+## 2018.12.17-20:07
+
++ Documentation finished on the code.
++ Test class created.
++ New public methods for traversing the BST.
++ New function for printing the tree to the console.
++ Removed TestCases.
+
 ## 2018.12.10-23:45
 
 + Added `contains` function.
 + Added `delete` function.
-+ Added `children` function to the `Node` class. to get ammount of children of the node.
++ Added `children` function to the `Node` class. to get amount of children of the node.
 
 ## 2018.11.08-23:52